+++ src/main/java/com/braintreegateway/PaymentInstrumentType.java
@@ -4,6 +4,7 @@ import com.braintreegateway.util.NodeWrapper;
 
 public final class PaymentInstrumentType {
     public static final String PAYPAL_ACCOUNT= "paypal_account";
+    public static final String PAYPAL_HERE = "paypal_here";
     @Deprecated public static final String EUROPE_BANK_ACCOUNT = "europe_bank_account";
     public static final String CREDIT_CARD = "credit_card";
     public static final String APPLE_PAY_CARD = "apple_pay_card";